9/1/2020 update
Form render now works, throws in both user fields and recipient fields to render. Need to update models to include some more things inside of it for the form. I also added a bootstrap theme for the index.html, the previous one is now named oldindex.html. Hopefully we can theme it this way